requirements.txt
absl-py==2.1.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
grpcio==1.68.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.2
pathlib==1.0.1
pillow==10.2.0
protobuf==5.29.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
typing_extensions==4.12.2
Werkzeug==3.1.3
